cruid_api/requirements.txt
asgiref==3.9.2
coverage==7.10.7
decouple==0.0.7
dj-database-url==2.1.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==21.2.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.6.0
